Moved CardDetailsModel to CardDetails feature folder
